feat(manage): action dev review comments
