+++ apps/manage/src/app/views/cases/view/application-updates/review/controller.test.js
@@ -631,106 +631,6 @@ describe('application updates review controller', () => {
 				}
 			});
 		});
-		it('should redirect to update details page if validation fails with empty string', async () => {
-			const mockDb = {
-				applicationUpdate: {
-					findUnique: mock.fn(() => ({
-						statusId: 'published'
-					}))
-				}
-			};
-			const mockReq = {
-				baseUrl: '/application-updates',
-				originalUrl: '/application-updates',
-				params: {
-					id: 'crown-dev-01',
-					updateId: 'app-update-01'
-				},
-				session: {
-					appUpdates: {
-						'app-update-01': {
-							details: 'an update',
-							publishNow: 'no'
-						}
-					}
-				},
-				body: {
-					details: ''
-				}
-			};
-			const mockRes = { render: mock.fn() };
-
-			const controller = buildSubmitUpdateDetails({ db: mockDb });
-			await controller(mockReq, mockRes);
-
-			assert.strictEqual(mockRes.render.mock.callCount(), 1);
-			assert.strictEqual(
-				mockRes.render.mock.calls[0].arguments[0],
-				'views/cases/view/application-updates/review/update-details.njk'
-			);
-			assert.deepStrictEqual(mockRes.render.mock.calls[0].arguments[1], {
-				applicationUpdateStatus: 'published',
-				details: '',
-				backLinkUrl: '/application-updates/app-update-01/review-published',
-				errors: {
-					details: {
-						msg: 'Enter update details'
-					}
-				},
-				errorSummary: [{ text: 'Enter update details', href: '#details' }]
-			});
-		});
-		it('should redirect to update details page if validation fails with string over 1000 chars', async () => {
-			const mockDb = {
-				applicationUpdate: {
-					findUnique: mock.fn(() => ({
-						statusId: 'published'
-					}))
-				}
-			};
-			const invalidLongString = 'hello'.repeat(201);
-			const mockReq = {
-				baseUrl: '/application-updates',
-				originalUrl: '/application-updates',
-				params: {
-					id: 'crown-dev-01',
-					updateId: 'app-update-01'
-				},
-				session: {
-					appUpdates: {
-						'app-update-01': {
-							details: 'an update',
-							publishNow: 'no'
-						}
-					}
-				},
-				body: {
-					details: invalidLongString
-				}
-			};
-
-			const mockRes = { render: mock.fn() };
-
-			const controller = buildSubmitUpdateDetails({ db: mockDb });
-			await controller(mockReq, mockRes);
-
-			assert.strictEqual(mockRes.render.mock.callCount(), 1);
-			assert.strictEqual(
-				mockRes.render.mock.calls[0].arguments[0],
-				'views/cases/view/application-updates/review/update-details.njk'
-			);
-			assert.deepStrictEqual(mockRes.render.mock.calls[0].arguments[1], {
-				applicationUpdateStatus: 'published',
-				details: invalidLongString,
-				backLinkUrl: '/application-updates/app-update-01/review-published',
-				errors: {
-					details: {
-						msg: 'Update details must be 1000 characters or less'
-					}
-				},
-				errorSummary: [{ text: 'Update details must be 1000 characters or less', href: '#details' }]
-			});
-		});
 	});
 	describe('buildPublishNowPage', () => {
 		it('should render publish now page with session value', async () => {
